Render engine 1.3.3 -Added multiple light support + preparations for shadowmapping

+++ src/mesh.cpp
@@ -46,7 +46,7 @@ void Mesh::Draw(Shader &shader)
         if(name == "texture_diffuse") number = std::to_string(diffuseNr++);
         else if(name == "texture_specular") number = std::to_string(specularNr++);
 
-        shader.setFloat(("material." + name + number).c_str(), i);
+        shader.setInt(("material." + name + number).c_str(), i);
         glBindTexture(GL_TEXTURE_2D, textures[i].id);
     }
     glActiveTexture(GL_TEXTURE0);
